rename: SelectColumn.js to SelectColumn.tsx
